+++ extensions/ql-vscode/src/interface.ts
@@ -30,7 +30,7 @@ import {
   RawResultsSortState,
 } from './pure/interface-types';
 import { Logger } from './logging';
-import { commandRunner } from './helpers';
+import { commandRunner } from './commandRunner';
 import * as messages from './pure/messages';
 import { CompletedQuery, interpretResults } from './query-results';
 import { QueryInfo, tmpDir } from './run-queries';
